+++ hpce5/src/include/housekeeping.hpp
@@ -15,7 +15,7 @@ constexpr uint64_t calculateImageSize(uint32_t w, uint32_t h, uint32_t bits) {
 }
 uint64_t calculateBufferSize(uint32_t w, uint32_t h, uint32_t bits,
                              uint32_t levels);
-uint64_t calculateChunkSize(uint32_t w, uint32_t h, uint32_t bits,
+uint32_t calculateChunkSize(uint32_t w, uint32_t h, uint32_t bits,
                             uint32_t levels, uint64_t bufferSize);
 
 uint8_t *allocateBuffer(uint64_t size, bool maximise = false);
